add base transaction class for DRY
