+++ app/services/transactions.py
@@ -3,16 +3,17 @@ from typing import Optional
 
 from app import models, schemas
 from app.logger import get_logger
-from app.services.accounts import AccountService
-from app.services.base import BaseService
-from app.utils.classes import RoundedDecimal
-from app.utils.exceptions import AccessDeniedError
-from app.utils.log_messages import ACCOUNT_USER_ID_MISMATCH
+from app.repository import Repository
+from app.services.base_transaction import BaseTransactionService
 
 logger = get_logger(__name__)
 
 
-class TransactionService(BaseService):
+class TransactionService(
+    BaseTransactionService,
+):
+    def __init__(self, repository: Optional[Repository] = None):
+        super().__init__(models.Transaction, repository)
 
     async def get_transaction_list(
         self,
@@ -21,34 +22,8 @@ class TransactionService(BaseService):
         date_start: datetime,
         date_end: datetime,
     ) -> list[models.Transaction]:
-        """
-        Retrieves a list of transactions within a specified period for a given account.
-
-        Args:
-            user: The user object.
-            account_id: The ID of the account.
-            date_start: The start date of the period.
-            date_end: The end date of the period.
-
-        Returns:
-            list[Transaction]: A list of transactions within the specified period.
-
-        Raises:
-            None
-        """
-
-        logger.info(
-            "Starting transaction list retrieval for user %s and account %s",
-            user.id,
-            account_id,
-        )
-        account = await self.repository.get(models.Account, account_id)
-
-        if account is None or account.user_id != user.id:
-            return []
-
-        return await self.repository.get_transactions_from_period(
-            account_id, date_start, date_end
+        return await super().get_transaction_list(
+            user, account_id, date_start, date_end
         )
 
     async def get_transaction(
@@ -68,26 +43,7 @@ class TransactionService(BaseService):
             None
         """
 
-        logger.info(
-            "Retrieving transaction with ID %s for user %s", transaction_id, user.id
-        )
-        transaction = await self.repository.get(models.Transaction, transaction_id)
-
-        if transaction is None:
-            logger.warning("Transaction with ID %s not found.", transaction_id)
-            return None
-
-        account = await self.repository.get(models.Account, transaction.account_id)
-
-        if account is None:
-            return None
-
-        if account.user_id == user.id:
-            logger.info("User ID verified. Returning transaction.")
-            return transaction
-
-        logger.warning(ACCOUNT_USER_ID_MISMATCH)
-        return None
+        return await super().get_transaction(user, transaction_id)
 
     async def create_transaction(
         self,
@@ -107,55 +63,7 @@ class TransactionService(BaseService):
         Raises:
             None
         """
-
-        logger.info("Creating new transaction for user %s", user.id)
-        account = await self.repository.get(models.Account, transaction_data.account_id)
-
-        if account is None:
-            return None
-
-        if not AccountService.has_user_access_to_account(user, account):
-            return None
-
-        db_transaction_information = models.TransactionInformation()
-        db_transaction_information.add_attributes_from_dict(
-            transaction_data.model_dump()
-        )
-
-        transaction = models.Transaction(
-            information=db_transaction_information,
-            account_id=account.id,
-            scheduled_transaction_id=transaction_data.scheduled_transaction_id,
-        )
-
-        if transaction_data.offset_account_id:
-            logger.info("Handling offset account for transaction.")
-            offset_transaction = await self._handle_offset_transaction(
-                user, transaction_data
-            )
-
-            if offset_transaction is None:
-                logger.warning(
-                    "User[id: %s] not allowed to access offset_account[id: %s]",
-                    user.id,
-                    transaction_data.offset_account_id,
-                )
-                raise AccessDeniedError(
-                    (
-                        f"User[id: {user.id}] not allowed to access "
-                        f"offset_account[id: {transaction_data.offset_account_id}]"
-                    )
-                )
-
-            transaction.offset_transaction = offset_transaction
-            offset_transaction.offset_transaction = transaction
-            await self.repository.save(offset_transaction)
-
-        account.balance += db_transaction_information.amount
-
-        await self.repository.save([account, transaction, db_transaction_information])
-
-        return transaction
+        return await super().create_transaction(user, transaction_data)
 
     async def _handle_offset_transaction(
         self,
@@ -175,39 +83,7 @@ class TransactionService(BaseService):
         Raises:
             None
         """
-
-        logger.info("Handling offset transaction for user %s", user.id)
-        offset_account_id = transaction_data.offset_account_id
-
-        if offset_account_id is None:
-            return None
-
-        offset_account = await self.repository.get(models.Account, offset_account_id)
-
-        if offset_account is None:
-            return None
-
-        if user.id != offset_account.user_id:
-            logger.warning("User ID does not match the offset account's User ID.")
-            return None
-
-        transaction_data.amount = RoundedDecimal(transaction_data.amount * -1)
-
-        offset_account.balance += transaction_data.amount
-
-        db_offset_transaction_information = models.TransactionInformation()
-        db_offset_transaction_information.add_attributes_from_dict(
-            transaction_data.model_dump()
-        )
-        offset_transaction = models.Transaction(
-            information=db_offset_transaction_information,
-            account_id=offset_account_id,
-            scheduled_transaction_id=transaction_data.scheduled_transaction_id,
-        )
-
-        await self.repository.save(offset_transaction)
-
-        return offset_transaction
+        return await super()._handle_offset_transaction(user, transaction_data)
 
     async def update_transaction(
         self,
@@ -229,65 +105,9 @@ class TransactionService(BaseService):
         Raises:
             None
         """
-
-        logger.info(
-            "Updating transaction with ID %s for user %s",
-            transaction_id,
-            current_user.id,
+        return await super().update_transaction(
+            current_user, transaction_id, transaction_information
         )
-        transaction = await self.repository.get(
-            models.Transaction,
-            transaction_id,
-            load_relationships_list=[models.Transaction.account],
-        )
-        if transaction is None:
-            return None
-
-        account = transaction.account
-
-        if current_user.id != account.user_id:
-            logger.warning(ACCOUNT_USER_ID_MISMATCH)
-            return None
-
-        # TODO: Use RoundedDecimal class here instead of round()
-        amount_updated = (
-            round(transaction_information.amount, 2) - transaction.information.amount
-        )
-
-        if transaction.offset_transactions_id:
-            logger.info("Handling offset transaction for update.")
-            offset_transaction = await self.repository.get(
-                models.Transaction,
-                transaction.offset_transactions_id,
-                load_relationships_list=[models.Transaction.account],
-            )
-            if offset_transaction is None:
-                return None
-
-            offset_account = offset_transaction.account
-
-            if offset_account is None or offset_account.user_id != current_user.id:
-                return None
-
-            offset_account.balance -= amount_updated
-            offset_transaction.information.amount = transaction_information.amount * -1
-
-        account_values = {"balance": account.balance + amount_updated}
-        await self.repository.update(models.Account, account.id, **account_values)
-
-        transaction_values = {
-            "amount": transaction_information.amount,
-            "reference": transaction_information.reference,
-            "date": transaction_information.date,
-            "category_id": transaction_information.category_id,
-        }
-        await self.repository.update(
-            models.TransactionInformation,
-            transaction.information.id,
-            **transaction_values,
-        )
-
-        return transaction
 
     async def delete_transaction(
         self, current_user: models.User, transaction_id: int
@@ -302,46 +122,4 @@ class TransactionService(BaseService):
         Returns:
             bool: True if the transaction is successfully deleted, False otherwise.
         """
-
-        logger.info(
-            "Deleting transaction with ID %s for user %s",
-            transaction_id,
-            current_user.id,
-        )
-        transaction = await self.repository.get(
-            models.Transaction,
-            transaction_id,
-            load_relationships_list=[models.Transaction.offset_transaction],
-        )
-
-        if transaction is None:
-            logger.warning("Transaction with ID %s not found.", transaction_id)
-            return None
-
-        account = await self.repository.get(models.Account, transaction.account_id)
-
-        if account is None:
-            return None
-
-        if current_user.id != account.user_id:
-            return None
-
-        amount = transaction.information.amount
-
-        if transaction.offset_transaction:
-            logger.info("Handling offset transaction for delete.")
-            offset_transaction = transaction.offset_transaction
-            offset_account = await self.repository.get(
-                models.Account, offset_transaction.account_id
-            )
-
-            if offset_account is None:
-                return None
-
-            offset_account.balance += amount
-            await self.repository.delete(transaction.offset_transaction)
-
-        account.balance -= amount
-        await self.repository.delete(transaction)
-
-        return True
+        return await super().delete_transaction(current_user, transaction_id)
